requirements.txt
psutil==5.9.0
requests>=2.25.1
python-dotenv==0.19.2